+++ client/src/components/common/CurrentUser.tsx
@@ -1,10 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import getData from 'api/fetch';
-import profileDefault from 'images/profile-default.png';
 import { useRecoilState } from 'recoil';
 import { chatWith } from 'recoil/store';
 // import { UserProps } from 'utils/types';
+
+import { defaultProfile } from 'images';
 
 interface UserProps {
   idx?: number;
@@ -31,7 +32,7 @@ const CurrentUser = () => {
       className="User"
       onClick={(): void => setChatWith(user)}
     >
-      <img src={profileDefault} />
+      <img src={defaultProfile} />
       {user}
     </CurrentUserBox>
   ));
