+++ client/src/images/icons/index.tsx
@@ -1,0 +1,32 @@
+import { ReactComponent as GnbHome } from 'images/icons/gnb-home.svg';
+import { ReactComponent as GnbGroup } from 'images/icons/gnb-group.svg';
+import { ReactComponent as GnbHomeActive } from 'images/icons/gnb-home-active.svg';
+import { ReactComponent as GnbGroupActive } from 'images/icons/gnb-group-active.svg';
+import { ReactComponent as LikeIcon } from 'images/icons/icon-like.svg';
+import { ReactComponent as CommentIcon } from 'images/icons/icon-comment.svg';
+import { ReactComponent as IconPublic } from 'images/icons/icon-public.svg';
+import { ReactComponent as IconPrivate } from 'images/icons/icon-private.svg';
+
+import gnbMessage from 'images/icons/gnb-message.svg';
+import gnbAlarm from 'images/icons/gnb-alarm.svg';
+import gnbSelector from 'images/icons/gnb-down-arrow.svg';
+import iconSearch from 'images/icons/icon-search.svg';
+import iconPhoto from 'images/icons/icon-photo.svg';
+import likeBadge from 'images/icons/icon-like-badge.svg';
+
+export {
+  GnbGroup,
+  GnbGroupActive,
+  GnbHome,
+  GnbHomeActive,
+  gnbMessage,
+  gnbAlarm,
+  gnbSelector,
+  iconPhoto,
+  iconSearch,
+  LikeIcon,
+  CommentIcon,
+  IconPublic,
+  IconPrivate,
+  likeBadge
+};
