+++ client/src/images/index.tsx
@@ -1,42 +1,6 @@
-import { ReactComponent as GnbHome } from 'images/gnb-home.svg';
-import { ReactComponent as GnbGroup } from 'images/gnb-group.svg';
-import { ReactComponent as GnbHomeActive } from 'images/gnb-home-active.svg';
-import { ReactComponent as GnbGroupActive } from 'images/gnb-group-active.svg';
-import gnbMessage from 'images/gnb-message.svg';
-import gnbAlarm from 'images/gnb-alarm.svg';
-import gnbSelector from 'images/gnb-down-arrow.svg';
-import profileDefault from 'images/profile-default.png';
-
-import mainLogo from 'images/main-logo.png';
-import iconSearch from 'images/icon-search.svg';
-
+// import profileDefault from 'images/profile-default.png';
 import defaultProfile from 'images/default-profile.jpg';
 import defaultCover from 'images/default-profile.jpg';
-import iconPhoto from 'images/icon-photo.svg';
+import mainLogo from 'images/main-logo.png';
 
-import { ReactComponent as LikeIcon } from 'images/icon-like.svg';
-import { ReactComponent as CommentIcon } from 'images/icon-comment.svg';
-import { ReactComponent as IconPublic } from 'images/icon-public.svg';
-import { ReactComponent as IconPrivate } from 'images/icon-private.svg';
-import likeBadge from 'images/icon-like-badge.svg';
-
-export {
-  GnbHome,
-  GnbGroup,
-  GnbHomeActive,
-  GnbGroupActive,
-  gnbMessage,
-  gnbAlarm,
-  gnbSelector,
-  profileDefault,
-  mainLogo,
-  defaultProfile,
-  defaultCover,
-  iconPhoto,
-  iconSearch,
-  LikeIcon,
-  CommentIcon,
-  IconPublic,
-  IconPrivate,
-  likeBadge
-};
+export { mainLogo, defaultProfile, defaultCover };
